edit clean

+++ data_cleaning.py
@@ -119,7 +119,7 @@ class DataCleaning():
         valid_timestamp_format = "%H:%M:%S"
         self.df['timestamp'] = pd.to_datetime(self.df['timestamp'], format=valid_timestamp_format, errors='coerce')
         valid_months = list(range(1, 13))
-        self.df = self.df[self.df['month'].astype(int).isin(valid_months)]
+        self.df = self.df[self.df['month'].isin(valid_months)]
         days_in_month = self.df.apply(lambda row: pd.Timestamp(row['year'], row['month'], 1).days_in_month, axis=1)
         self.df = self.df[self.df['day'].between(1, days_in_month)]
         self.df['year'] = pd.to_numeric(self.df['year'], errors='coerce')
